chatbot-llm-main/requirements.txt
Flask==3.0.0
google-cloud-texttospeech==2.18.0
google-cloud-speech==2.22.0
groq~=0.13.1
python-dotenv~=1.0.1

gTTS~=2.5.4